Don't require top-level documentation comments
Because of how easy it is for documentation blocks to get out-of-date, I
prefer to document the behaviour of code in specs and any non-obvious
implementation details in commit messages.

I'm therefore disabling the `Style/Documentation` cop [1] to avoid
having to add relatively low value top-level documentation blocks to
boilerplate code.

[1] https://docs.rubocop.org/rubocop/cops_style.html#styledocumentation