Add users table and model
Rail's built-in uniqueness validation [1] is prone to race conditions so
I've decided to include a unique key constraint in the database schema
to guarantee the uniqueness of user email addresses instead. This does,
however, mean that we'll need to handle `ActiveRecord::RecordNotUnique`
exceptions to give the user meaningful feedback when attempting to
register a user with a duplicate email address.

I had to install the BCrypt gem to hash the user password using Rails'
built-in `has_secure_password` method [2]. As well as hashing that
method also takes care of enforcing presence validation on the
`password` attribute.

[1] https://api.rubyonrails.org/v7.0.4/classes/ActiveRecord/Validations/ClassMethods.html#method-i-validates_uniqueness_of
[2] https://api.rubyonrails.org/v7.0.4/classes/ActiveModel/SecurePassword/ClassMethods.html#method-i-has_secure_password