Redirect already registered users to log in
We have a unique key constraint in the database to ensure that no two
user accounts have the same email address. When this constraint is
violated ActiveRecord raises an `ActiveRecord::RecordNotUnique` error
which was previously going unhandled.

Now that we have a log in page, we can rescue the error and redirect the
user to log in instead.

We're letting the error occur and then rescuing it, because the
alternative of checking for an existing user upfront is prone to race
conditions; especially if a user with JavaScript disabled rapidly double
clicks on the submit button, for example.

In theory a duplicate email address is currently the only scenario in
which an `ActiveRecord::RecordNotUnique` error could be raised here.
However that may cease to be true if more columns with unique
constraints are added later so I'm also double checking that there is
indeed another user with the same email address before redirecting. If
there isn't another user with the same email address, then the redirect
and alert message would be very misleading and so it would be better to
respond with the default server error instead.