Add questions table and model
I've made the `author_id` column nullable with the a foreign key
constraint that nullifies on deletion in anticipation that users will
eventually be able to delete their accounts.

When that happens we'll want the questions to remain for the benefit of
other users, but it'll no longer be attributable to a known author.