Add view helper for displaying the name nilable user
The `#author` attribute on questions is deliberately nilable to allow
for user deletion. I expect the same will also be true of user
attributes on other models such as answers.

This view helper will be used to either display the name of the user if
possible, or 'an anonymous user' if the user is nil.