Add question view
Due to the time constraints on this exercise this is currently a very
basic view.

I'm using Rail's built in `time_ago_in_words` view helper [1] to display
the age of the question in a way that makes sense regardless of the
user's time zone. In a real application I'd also want to display the
absolute creation time (maybe in a popover) in the user's current time
zone.

I'm also using the `simple_format` helper [2] to ensure that line breaks
in the question body are rendered. In a real application I would expect
to provide richer formatting options through something like markdown or
even a full rich text editor.

[1] https://api.rubyonrails.org/v7.0.4/classes/ActionView/Helpers/DateHelper.html#method-i-time_ago_in_words
[2] https://api.rubyonrails.org/v7.0.4/classes/ActionView/Helpers/TextHelper.html#method-i-simple_format