Add form for posting an answer
I've reached the time limit for this exercise so I've had to cut some
more corners here.

In particular this create action makes no effort to gracefully handle
validation errors. If the answer can't be persisted then the user will
see a server error. Fortunately the `required` attribute on the field
should cause modern browsers to prevent the form from being submitted
without an answer body, which is the only validation on the field.

I also ran out of time to redirect unauthenticated users to log in.